+++ src/main/java/subway/domain/station/StationName.java
@@ -1,21 +1,21 @@
-package subway.domain;
+package subway.domain.station;
 
 import java.util.Objects;
 
-public class LineName {
+public class StationName {
     private static final int MAX_NAME_LENGTH = 10;
     private static final int MIN_NAME_LENGTH = 3;
 
     private final String name;
 
-    public LineName(final String name) {
+    public StationName(final String name) {
         validateLength(name);
         this.name = name;
     }
 
     private void validateLength(final String name) {
         if (name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("노선 이름 길이는 " + MIN_NAME_LENGTH + "~" + MAX_NAME_LENGTH + "여야 합니다.");
+            throw new IllegalArgumentException("역 이름 길이는 " + MIN_NAME_LENGTH + "~" + MAX_NAME_LENGTH + "여야 합니다.");
         }
     }
 
@@ -27,7 +27,7 @@ public class LineName {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        final LineName lineName = (LineName) o;
+        final StationName lineName = (StationName) o;
         return Objects.equals(name, lineName.name);
     }
