+++ src/main/java/subway/domain/station/Stations.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import java.util.List;
 import java.util.stream.Collectors;
