+++ src/test/java/subway/domain/line/LineNameTest.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.domain.line.LineName;
 
 class LineNameTest {
