+++ src/test/java/subway/domain/station/StationNameTest.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.domain.station.StationName;
 
 class StationNameTest {
